fix: replace CSP hash with unsafe-inline for anchor-js styles
In Safari only, the CSP hash for AnchorJS styles wasn't working. This
prevented AnchorJS from running and broke the navigation in Safari.

My best explanation was that this is a browser bug (see details here:
https://github.com/bryanbraun/checkboxland/issues/15 ).

I explored a lot of solutions but only unsafe-inline seemed to work.
I couldn't use a nonce since there was no server component to this
page. It's unclear if Safari was looking for a different hash value.

Ultimately, I'm ok with this for now.